+++ zebrad/tests/common/launch.rs
@@ -210,7 +210,7 @@ where
 /// Spawns a zebrad instance on `network` to test lightwalletd with `test_type`.
 ///
 /// If `use_internet_connection` is `false` then spawn, but without any peers.
-/// This prevents it from downloading blocks. Instead, use the `ZEBRA_CACHE_DIR`
+/// This prevents it from downloading blocks. Instead, use the `STATE_CACHE_DIR`
 /// environmental variable to provide an initial state to the zebrad instance.
 ///
 /// Returns:
@@ -261,7 +261,7 @@ pub fn spawn_zebrad_for_rpc<S: AsRef<str> + Debug>(
 /// Otherwise, just create an empty state in this test's new temporary directory.
 ///
 /// If `use_internet_connection` is `false` then spawn, but without any peers.
-/// This prevents it from downloading blocks. Instead, use the `ZEBRA_CACHE_DIR`
+/// This prevents it from downloading blocks. Instead, use the `STATE_CACHE_DIR`
 /// environmental variable to provide an initial state to the zebrad instance.
 ///
 /// Returns: